refactor(tui): extract UI state, selection, and node snapshot types
Chord and modal state, per-pane view state, intents, and text selection
move into tui::state; clipboard handling into tui::selection; the
node-facing snapshot types into tui::snapshot; and the resize-drag types
join the geometry they describe.
